packet: move helpers which rewrite headers to a separate file
Move helpers which rewrite headers, such as vlan_tag_pop or swap_src_dst_ipv4
to a separate header file so that they can be reused.
